hacking to fixup authenticate
was getting a weird error.  turns out the problem (I think) was that
the alpine version of docker postgresql has issues with truncating
tables.  It could not get rid of the duplicate unique id for the test
account email.

Whatever.  When I switch to postgres:latest version with pg_tap
installed, no problems